+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.19**
+* action method parameter cache #1054
 * Route information missed in API doc for dynamic aliases #1052
 * command `route.print -t` triggers error #1053
 * `@WsEndpoint` annotation now failed to register route #1050